refactor files layout
