+++ NetRocks/src/Protocol/SSH/ProtocolSCP.h
@@ -1,25 +1,18 @@
 #pragma once
 #include <memory>
-#include <string>
-#include <map>
-#include <StringConfig.h>
-#include "Protocol.h"
+#include "../Protocol.h"
 
-class ProtocolSMB : public IProtocol, public std::enable_shared_from_this<ProtocolSMB>
+struct SSHConnection;
+
+class ProtocolSCP : public IProtocol
 {
-	std::string _host;
+	std::shared_ptr<SSHConnection> _conn;
+	struct timespec _now{};
 
 public:
-	StringConfig _protocol_options;
-	std::map<std::string, FileInformation> _cached_net;
-
-	ProtocolSMB(const std::string &host, unsigned int port,
-		const std::string &username, const std::string &password, const std::string &protocol_options) throw (std::runtime_error);
-	virtual ~ProtocolSMB();
-
-	const std::vector<std::string> &EnumHosts();
-
-	std::string RootedPath(const std::string &path);
+	ProtocolSCP(const std::string &host, unsigned int port, const std::string &username,
+		const std::string &password, const std::string &options) throw (std::runtime_error);
+	virtual ~ProtocolSCP();
 
 	virtual mode_t GetMode(const std::string &path, bool follow_symlink = true) throw (std::runtime_error);
 	virtual unsigned long long GetSize(const std::string &path, bool follow_symlink = true) throw (std::runtime_error);
@@ -37,7 +30,10 @@ public:
 	virtual void SymlinkCreate(const std::string &link_path, const std::string &link_target) throw (std::runtime_error);
 	virtual void SymlinkQuery(const std::string &link_path, std::string &link_target) throw (std::runtime_error);
 
+
 	virtual std::shared_ptr<IDirectoryEnumer> DirectoryEnum(const std::string &path) throw (std::runtime_error);
 	virtual std::shared_ptr<IFileReader> FileGet(const std::string &path, unsigned long long resume_pos = 0) throw (std::runtime_error);
 	virtual std::shared_ptr<IFileWriter> FilePut(const std::string &path, mode_t mode, unsigned long long size_hint, unsigned long long resume_pos = 0) throw (std::runtime_error);
+
+	virtual void ExecuteCommand(const std::string &working_dir, const std::string &command_line, const std::string &fifo) throw (std::runtime_error);
 };
